backend/requirements.txt
Flask==2.0.3
psycopg2-binary==2.9.1
Werkzeug==2.0.1
gunicorn==20.1.0
requests==2.26.0
python-dotenv==1.0.0
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.3.24
